Replace react-loadable with react.lazy
