+++ package.json
@@ -123,7 +123,6 @@
     "react-collapsible": "^2.6.0",
     "react-dom": "^16.10.2",
     "react-feather": "^2.0.3",
-    "react-loadable": "^5.5.0",
     "react-markdown": "^4.2.2",
     "react-piwik": "^1.6.0",
     "react-rangeslider": "^2.2.0",
