+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { PureComponent, Suspense, lazy } from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 import { Route, Switch, Router, Redirect } from 'react-router-dom';
@@ -10,9 +10,6 @@ import GlobalStyles from '~/styles/Global';
 import Menu from '~/components/Menu';
 import PrivateRoute from '~/components/PrivateRoute';
 import Home from '~/pages/Home';
-import Markdown from '~/pages/Markdown';
-import MapView from '~/pages/Map';
-import Analysis from '~/pages/Analysis';
 import Login from '~/pages/User/pages/Login';
 import Signup from '~/pages/User/pages/Signup';
 import Profile from '~/pages/User/pages/Profile';
@@ -20,8 +17,12 @@ import ForgotPassword from '~/pages/User/pages/ForgotPassword';
 import ResetPassword from '~/pages/User/pages/ResetPassword';
 import UserVerify from '~/pages/User/pages/Verify';
 import { verify } from '~/pages/User/UserState';
-import Reports from '~/pages/Reports';
-import KatasterKI from '~/pages/KatasterKI';
+
+const Analysis = lazy(() => import('~/pages/Analysis'));
+const KatasterKI = lazy(() => import('~/pages/KatasterKI'));
+const MapView = lazy(() => import('~/pages/Map'));
+const Markdown = lazy(() => import('~/pages/Markdown'));
+const Reports = lazy(() => import('~/pages/Reports'));
 
 const AppContent = styled.div`
   width: 100%;
@@ -32,6 +33,8 @@ const AppWrapper = styled.div`
   height: 100%;
   position: relative;
 `;
+
+const Loading = () => <div>Loading...</div>;
 
 class App extends PureComponent {
   componentDidMount() {
@@ -49,75 +52,79 @@ class App extends PureComponent {
             <AppWrapper>
               {!isEmbedMode && <Menu />}
               <AppContent>
-                <Switch>
-                  <Route exact path="/" component={Home} />
+                <Suspense fallback={<Loading />}>
+                  <Switch>
+                    <Route exact path="/" component={Home} />
 
-                  {/* standard markdown pages */
-                  config.staticpages.map((page) => (
+                    {/* standard markdown pages */
+                    config.staticpages.map((page) => (
+                      <Route
+                        key={page}
+                        path={page.route}
+                        render={() => <Markdown page={page.key} />}
+                      />
+                    ))}
+
+                    {/* user pages */}
+                    <Route path={config.routes.signup} component={Signup} />
+                    <Route path={config.routes.login} component={Login} />
                     <Route
-                      key={page}
-                      path={page.route}
-                      render={() => <Markdown page={page.key} />}
+                      path={config.routes.forgotPassword}
+                      component={ForgotPassword}
                     />
-                  ))}
+                    <Route
+                      path={`${config.routes.resetPassword}/:uid/:token`}
+                      component={ResetPassword}
+                    />
+                    <Route
+                      path={`${config.routes.userVerify}/:uid/:token`}
+                      component={UserVerify}
+                    />
+                    <PrivateRoute
+                      path={config.routes.profile}
+                      token={token}
+                      component={Profile}
+                    />
 
-                  {/* user pages */}
-                  <Route path={config.routes.signup} component={Signup} />
-                  <Route path={config.routes.login} component={Login} />
-                  <Route
-                    path={config.routes.forgotPassword}
-                    component={ForgotPassword}
-                  />
-                  <Route
-                    path={`${config.routes.resetPassword}/:uid/:token`}
-                    component={ResetPassword}
-                  />
-                  <Route
-                    path={`${config.routes.userVerify}/:uid/:token`}
-                    component={UserVerify}
-                  />
-                  <PrivateRoute
-                    path={config.routes.profile}
-                    token={token}
-                    component={Profile}
-                  />
+                    {/* map pages */}
+                    <Route
+                      path={`(${config.routes.status}|${config.routes.projects}|/my-hbi)`}
+                      component={MapView}
+                    />
 
-                  {/* map pages */}
-                  <Route
-                    path={`(${config.routes.status}|${config.routes.projects}|/my-hbi)`}
-                    component={MapView}
-                  />
+                    {/* reports page */}
+                    <Route
+                      path={`${config.routes.reports.index}`}
+                      component={Reports}
+                    />
 
-                  {/* reports page */}
-                  <Route
-                    path={`${config.routes.reports.index}`}
-                    component={Reports}
-                  />
+                    {/* kataster survey page */}
+                    <Route
+                      path={config.routes.katasterKI.landing}
+                      component={KatasterKI}
+                    />
 
-                  {/* kataster survey page */}
-                  <Route
-                    path={config.routes.katasterKI.landing}
-                    component={KatasterKI}
-                  />
+                    {/* analysis pages */}
+                    <Route
+                      path={`${config.routes.analyse}/planungen/:districtName?`}
+                      component={Analysis}
+                    />
 
-                  {/* analysis pages */}
-                  <Route
-                    path={`${config.routes.analyse}/planungen/:districtName?`}
-                    component={Analysis}
-                  />
+                    {/* reports page */}
+                    <Route
+                      exact
+                      path={
+                        config.routes.reports
+                          .temporarily_forward_from_this_to_index
+                      }
+                      render={() => (
+                        <Redirect to={config.routes.reports.index} />
+                      )}
+                    />
 
-                  {/* reports page */}
-                  <Route
-                    exact
-                    path={
-                      config.routes.reports
-                        .temporarily_forward_from_this_to_index
-                    }
-                    render={() => <Redirect to={config.routes.reports.index} />}
-                  />
-
-                  <Route render={() => <Markdown page="nomatch" />} />
-                </Switch>
+                    <Route render={() => <Markdown page="nomatch" />} />
+                  </Switch>
+                </Suspense>
               </AppContent>
             </AppWrapper>
           </LastLocationProvider>
