--- src/pages/Markdown/Markdown.js
@@ -1,40 +1,0 @@
-import React, { PureComponent } from 'react';
-import ky from 'ky';
-
-import ContentPageWrapper from '~/components/ContentPageWrapper';
-import MarkdownContent from '~/pages/Markdown/components/MarkdownContent';
-
-class MarkdownPage extends PureComponent {
-  constructor(props) {
-    super(props);
-    this.state = {
-      content: false
-    };
-  }
-
-  componentDidMount() {
-    this.loadContent();
-  }
-
-  componentDidUpdate(prevProps) {
-    if (prevProps.page !== this.props.page) {
-      this.loadContent();
-    }
-  }
-
-  async loadContent() {
-    const { page } = this.props;
-    const content = await ky.get(`/markdown/${page}.md`).text();
-    return this.setState({ content });
-  }
-
-  render() {
-    return (
-      <ContentPageWrapper>
-        <MarkdownContent content={this.state.content} />
-      </ContentPageWrapper>
-    );
-  }
-}
-
-export default MarkdownPage;
