+++ src/pages/Reports/pages/SubmitReport/ReportDialog.js
@@ -12,7 +12,6 @@ import { actions as submitReportStateActions } from '~/pages/Reports/state/Submi
 
 import config from '~/pages/Reports/config';
 import OverviewMapNavBar from '~/pages/Reports/pages/OverviewMap/components/OverviewMapNavBar';
-import Markdown from '~/pages/Markdown/Markdown';
 import LocateModeChooser from './pages/LocateModeChooser';
 import LocateMeMap from './components/LocateMeMap';
 import BikestandsForm from './pages/BikestandsForm';
@@ -21,6 +20,8 @@ import FormProgressBar from '~/pages/Reports/pages/SubmitReport/components/FormP
 import ReportSubmitted from './pages/ReportSubmitted';
 import BicycleParkingGarageForm from './pages/BicycleParkingGarageForm';
 import ThanksPage from './pages/ThanksPage';
+
+const Markdown = React.lazy(() => import('~/pages/Markdown'));
 
 const LoaderWrapper = styled.div`
   width: 100%;
