--- src/pages/User/pages/EmailVerification/index.js
@@ -1,7 +1,0 @@
-import Loadable from 'react-loadable';
-
-export default Loadable({
-  loader: () =>
-    import(/* webpackChunkName: "EmailVerification" */ './EmailVerification'),
-  loading: () => null
-});
