--- src/pages/User/pages/ForgotPassword/index.js
@@ -1,7 +1,0 @@
-import Loadable from 'react-loadable';
-
-export default Loadable({
-  loader: () =>
-    import(/* webpackChunkName: "PasswordReset" */ './ForgotPassword'),
-  loading: () => null
-});
