--- src/pages/User/pages/Login/index.js
@@ -1,6 +1,0 @@
-import Loadable from 'react-loadable';
-
-export default Loadable({
-  loader: () => import(/* webpackChunkName: "Login" */ './Login'),
-  loading: () => null
-});
