--- src/pages/User/pages/ResetPassword/index.js
@@ -1,7 +1,0 @@
-import Loadable from 'react-loadable';
-
-export default Loadable({
-  loader: () =>
-    import(/* webpackChunkName: "PasswordReset" */ './ResetPassword'),
-  loading: () => null
-});
