--- src/pages/User/pages/Signup/index.js
@@ -1,6 +1,0 @@
-import Loadable from 'react-loadable';
-
-export default Loadable({
-  loader: () => import(/* webpackChunkName: "Signup" */ './Signup'),
-  loading: () => null
-});
